JAMES-3715 Rename: UpstreamHandler -> InboundHandler
